% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks_analysis.R
\name{calc_metric_stability}
\alias{calc_metric_stability}
\title{Calculate Metric Stability
Calculates 3 major network features from edgelists of TFs and gene targets in igraph.rds format or txt files. Edgelists need to be in tab separated format with TF in column 1, Gene in column 2. Additional columns are ignored in the current implementation.}
\usage{
calc_metric_stability(indir, outfile = NULL, fraction,
  metric = c("outdegree", "indegree", "betweenness"), verbose = FALSE,
  in_ext = "_graph.rds", out_ext = ".rds")
}
\arguments{
\item{indir}{Directory containing edgelists for all networks. MUST BE IN .rds for igraph objects or tab separated .txt files. It helps if the the filenames include the sample ID followed by the file extension.}

\item{outfile}{File to write final metric matrix. Can be NULL to return matrix in session}

\item{fraction}{Fraction of edges in a given network to remove for evaluating stability}

\item{metric}{Metric to calculate}

\item{verbose}{Suppress warnings and output logging}

\item{in_ext}{File extension of edgelists}

\item{out_ext}{File extension of output file}
}
\description{
Calculate Metric Stability
Calculates 3 major network features from edgelists of TFs and gene targets in igraph.rds format or txt files. Edgelists need to be in tab separated format with TF in column 1, Gene in column 2. Additional columns are ignored in the current implementation.
}
