% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks_analysis.R
\name{calc_TF_score}
\alias{calc_TF_score}
\title{Calculate TF-Score
Calculates TF Score for a set of matrices where TFs are rows and samples/cell lines are columns. This works best for RANKED matrices. Only keeps TFs present in all provided matrices.}
\usage{
calc_TF_score(outdegree_df, expression_df, betweenness_df = NULL,
  method = c("simple", "complex"))
}
\arguments{
\item{outdegree_df}{Matrix of outdegree centrality for TFs}

\item{expression_df}{Matrix of expression for TFs}

\item{betweenness_df}{Matrix of betweenness centrality for TFs is optional}

\item{method}{Simple average of the matrices or average + standard deviation.}
}
\description{
Calculate TF-Score
Calculates TF Score for a set of matrices where TFs are rows and samples/cell lines are columns. This works best for RANKED matrices. Only keeps TFs present in all provided matrices.
}
