% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks_analysis.R
\name{get_regulatory_targets}
\alias{get_regulatory_targets}
\title{Get Target Genes
Identifies target genes of a given TF}
\usage{
get_regulatory_targets(graph, TF)
}
\arguments{
\item{graph}{igraph object for a given sample/cell line}

\item{TF}{TF of interest}
}
\description{
Get Target Genes
Identifies target genes of a given TF
}
