% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks_analysis.R
\name{rank_mat}
\alias{rank_mat}
\title{Rank Matrices
Ranks a matrix by either row or column such that high values in original matrix are given rank 1-n}
\usage{
rank_mat(matrix, margin = "column")
}
\arguments{
\item{matrix}{Matrix of whatever you want to rank, must be numeric}

\item{margin}{How to rank matrix, row calculates the rank within rows, column calculates the rank within columns}
}
\description{
Rank Matrices
Ranks a matrix by either row or column such that high values in original matrix are given rank 1-n
}
